+++ src/App/MyRange/Store/Article/Article.container.jsx
@@ -1,7 +1,7 @@
-import { BounceBall, LoadBar, } from 'Common/components/Loaders';
 import React, { PureComponent, } from 'react';
 
 import ArticleActions from './components/ArticleActions';
+import { LoadBar, } from 'Common/components/Loaders';
 import MiniCard from 'Common/components/MiniCard';
 import PlanogramDetails from './components/PlanogramDetails';
 import SlideWrap from 'Common/styles/SlideWrap';
@@ -9,151 +9,104 @@ import { getPath, } from 'utilibelt';
 import getStatus from './getStatus';
 import gqlMyRangeStoreArticle from './Article.gql';
 import noProductImage from 'Assets/noProductImage.webp';
-import { toast, } from 'react-toastify';
 import { withRouter, } from 'react-router-dom';
 
 @SlideWrap
 @withRouter
 @gqlMyRangeStoreArticle
 class MyRangeStoreArticle extends PureComponent {
-  constructor ( props ) {
-    super( props );
-
-    this.state = {
-      gettingBig : false,
-      swipe      : 0,
-    };
-  }
-
   componentDidMount () {
     const { item, } = this.props;
 
     !item && this.setTitle();
   }
 
-  componentDidUpdate ( prevProps ) {
+  componentDidUpdate () {
     const {
-      article: { error, } = {},
+      error,
       history: { push, },
       location,
-      match: { params = {}, },
+      match: { params, },
     } = this.props;
-    const { article: { error: pastError, } = {}, } = prevProps;
 
-    if ( error !== pastError && error && params._id ) {
-      toast.error( 'The provided article number does not exist' );
-
-      push( location.pathname.replace( `/${params._id}`, '' ) );
-    }
-  }
-
-  componentWillUnmount () {
-    const { timer, } = this.state;
-
-    if ( timer ) clearTimeout( timer );
+    if ( params._id && error ) push( location.pathname.replace( `/${params._id}`, '' ) );
   }
 
   setTitle () {
-    const { article: { getArticle: article, }, } = this.props;
+    const { article = {}, store = {}, } = this.props;
 
-    document.title = `myBWS ${article.id} - ${article.description} Range Details`;
+    if ( article.id && store.id ) document.title = `myRange | ${store.id} - ${store.name} | ${article.id} - ${article.description}`;
   }
 
+  handleOpenBig = () => {
+    const { history, loading, article, match, } = this.props;
+    const isBig = !!match.params._id;
+
+    if ( !loading && !isBig ) history.push( `${match.url}/${article.id}` );
+  };
+
+  handleCloseBig = () => {
+    const { unmountMe, history, match, } = this.props;
+
+    unmountMe( () => history.push( match.url.replace( `/${match.params._id}`, '' ) ) );
+  };
+
   render () {
-    const {
-      article: {
-        getArticle: article = {},
-        getUser: { role, } = {},
-        getStore: store = {},
-        getPlanograms: planograms = [],
-        loading = !article.id,
-      } = {},
-      match = {},
-      history = {},
-      swipe,
-      unmountMe,
-      style,
-    } = this.props;
+    const { article, user, store, planograms, loading, match, style, } = this.props;
+    const isBig = !!match.params._id;
+
+    if ( loading ) {
+      return (
+        <MiniCard
+          isBig={ isBig } loading={ loading } style={ style }
+          top>
+          <MiniCard.Header description='' ribbon='Loading...' title='' />
+          <MiniCard.Body isBig={ isBig }>
+            <LoadBar />
+          </MiniCard.Body>
+        </MiniCard>
+      );
+    }
+    const { role, } = user;
     const { organisation, district, } = store;
     const orgdist = `${organisation}.${district}`;
     const { status, supply, isBlocked, } = getStatus( article, planograms, store );
 
     if ( isBlocked ) article.blocked = isBlocked;
-    const isBig = !!match.params._id;
-    const altUrl = `https://edgmedia.bws.com.au/bws/media/products/${article.id}-1.png?impolicy=Prod_`;
-    const imgUrl = `https://media.danmurphys.com.au/dmo/product/${article.id}-1.png?impolicy=PROD_`;
+    const imgUrl = `https://media.danmurphys.com.au/dmo/product/${article.id}-1.png?impolicy=PROD_RETINA_SM`;
+    const ribbonColor = [
+      'Pog Range',
+      'Promo/Season',
+      'Customer 1st',
+    ].includes( status )
+      ? '#eda12a'
+      : status === 'Available'
+        ? '#29dc3f'
+        : '#f34859';
 
     return (
-      <MiniCard
-        isBig={ isBig }
-        loading={ loading }
-        style={
-          isBig && swipe
-            ? {
-              ...style,
-              transform: `translateY(-${swipe}px)`,
-            }
-            : style
-        }
-        top>
+      <MiniCard isBig={ isBig } style={ style } top>
         <MiniCard.Header
-          alt={ `${altUrl}Retina_SM` }
-          avatar={ loading ? null : `${imgUrl}RETINA_SM` }
+          avatar={ imgUrl }
           data-for='global'
-          description={
-            !loading &&
-            `${article.id}
-          ${article.group}`
-          }
+          description={ `${article.id || ''}\n${article.group || ''}` }
           noimg={ noProductImage }
-          onClick={ () => !loading && !isBig && history.push( `${match.url}/${article.id}` ) }
-          onCloseClick={
-            isBig
-              ? () => {
-                unmountMe( () => history.push( match.url.replace( `/${match.params._id}`, '' ) ) );
-              }
-              : null
-          }
-          placeholder={ loading ? null : `${imgUrl}XS` }
-          ribbon={ !loading && ( status || 'Loading...' ) }
-          ribbonColor={
-            [
-              'Pog Range',
-              'Promo/Season',
-              'Customer 1st',
-            ].includes( status )
-              ? '#eda12a'
-              : status === 'Available'
-                ? '#29dc3f'
-                : '#f34859'
-          }
-          stamp={
-            !loading &&
-            `$${getPath( article,
-                         `${orgdist}.price.mpk`,
-                         getPath( article, `${orgdist}.price.ea`, getPath( article, `${orgdist}.price.car`, ' - ' ) ) )}`
-          }
+          onClick={ this.handleOpenBig }
+          onCloseClick={ isBig && this.handleCloseBig }
+          placeholder={ `${imgUrl}XS` }
+          ribbon={ status }
+          ribbonColor={ ribbonColor }
+          stamp={ `$${getPath( article,
+                               `${orgdist}.price.mpk`,
+                               getPath( article, `${orgdist}.price.ea`, getPath( article, `${orgdist}.price.car`, ' - ' ) ) )}` }
           title={ article.description || '' }>
-          {isBig &&
-            ( planograms && article.description
-              ? <PlanogramDetails
-                article={ article } orgdist={ orgdist } planograms={ planograms }
-                supply={ supply } />
-              : (
-                <div
-                  style={ {
-                    bottom   : '15%',
-                    left     : '15%',
-                    position : 'absolute',
-                  } }>
-                  <BounceBall />
-                  <span style={ { marginLeft: 16, } }>Loading additional info...</span>
-                </div>
-              ) )}
+          {isBig && <PlanogramDetails
+            article={ article } orgdist={ orgdist } planograms={ planograms }
+            supply={ supply } />}
         </MiniCard.Header>
         <MiniCard.Body
           isBig={ isBig }
-          message={ !loading && ArticleActions }
+          message={ ArticleActions }
           messageProps={ {
             isBig,
             item    : article,